ittaita

+++ app/Http/Controllers/TasksController.php
@@ -11,6 +11,24 @@ class TasksController extends Controller
     // getでmessages/にアクセスされた場合の「一覧表示処理」
     public function index()
     {
+        $data = [];
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            // ユーザの投稿の一覧を作成日時の降順で取得
+            // （後のChapterで他ユーザの投稿も取得するように変更しますが、現時点ではこのユーザの投稿のみ取得します）
+            
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+        }
+
+        // Welcomeビューでそれらを表示
+        return view('welcome', $data);
+        
         // メッセージ一覧を取得
         $tasks = Task::all();
 
@@ -45,16 +63,30 @@ class TasksController extends Controller
     {
         // バリデーション
         $request->validate([
+       
             'status' => 'required|max:10',   // 追加
             'content' => 'required|max:255',
         ]);
         
         // メッセージを作成
-        $task = new Task;
-        $task->status = $request->status;   // 追加
-        $task->content = $request->content;
-        $task->save();
-
+        // Task::create([
+        //     'user_id' => \Auth::user()->id,
+        //     'status' => $date['status'],
+        //     'content' => $date['content'],
+        // ]);
+        
+        // $task = new Task;
+        // $task->user_id = $request->user()->id;
+        // $task->status = $request->status;   // 追加
+        // $task->content = $request->content;
+        // $task->save();
+        
+         // 認証済みユーザ（閲覧者）の投稿として作成（リクエストされた値をもとに作成）
+        $request->user()->tasks()->create([
+            'status' => $request->status,
+            'content' => $request->content,
+        ]);
+        
         // トップページへリダイレクトさせる
         return redirect('/');
 
@@ -72,9 +104,19 @@ class TasksController extends Controller
         $task = Task::findOrFail($id);
 
         // メッセージ詳細ビューでそれを表示
+        if (\Auth::id() == $task->user_id) {
         return view('tasks.show', [
             'task' => $task,
         ]);
+        }
+        
+         // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は、、、
+        // if (\Auth::id() == $task->user_id) {
+            // $task->show();
+        // }
+
+        // トップページへリダイレクトさせる
+        return redirect('/');
 
     }
 
@@ -90,9 +132,14 @@ class TasksController extends Controller
         $task = Task::findOrFail($id);
 
         // メッセージ編集ビューでそれを表示
+        if (\Auth::id() == $task->user_id) {
         return view('tasks.edit', [
             'task' => $task,
         ]);
+        }
+        
+        // トップページへリダイレクトさせる
+        return redirect('/');
     }
 
     /**
@@ -106,6 +153,7 @@ class TasksController extends Controller
     {
         // バリデーション
         $request->validate([
+            
             'status' => 'required|max:10',   // 追加
             'content' => 'required|max:255',
         ]);
@@ -113,9 +161,14 @@ class TasksController extends Controller
         // idの値でメッセージを検索して取得
         $task = Task::findOrFail($id);
         // メッセージを更新
+        $task->user_id = $request->user()->id;
         $task->status = $request->status;   // 追加
         $task->content = $request->content;
         $task->save();
+        
+        if (\Auth::id() == $task->user_id) {
+            $task->update();
+        }
 
         // トップページへリダイレクトさせる
         return redirect('/');
@@ -130,9 +183,12 @@ class TasksController extends Controller
     public function destroy($id)
     {
         // idの値でメッセージを検索して取得
-        $task = Task::findOrFail($id);
-        // メッセージを削除
-        $task->delete();
+        $task = \App\Task::findOrFail($id);
+ //dd(\Auth::id() . '|' . $task->user_id);       
+        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は、投稿を削除
+        if (\Auth::id() == $task->user_id) {
+            $task->delete();
+        }
 
         // トップページへリダイレクトさせる
         return redirect('/');
